+++ quantforge/__init__.py
@@ -694,6 +694,7 @@ from .expint import e1, ei, en
 from .lambert import lambert_w0, lambert_wm1
 from .elliptic import agm, elliptic_k, elliptic_e
 from .bessel import bessel_j0, bessel_j1, bessel_jn, bessel_y0, bessel_y1
+from .fresnel import fresnel_c, fresnel_s, dawson
 from .dual import Dual, derivative as dual_derivative
 from .dual_calculus import dual_gradient, dual_newton
 from .hyperdual import HyperDual, derivatives as hyperdual_derivatives, second_derivative
@@ -1398,7 +1399,7 @@ try:  # pragma: no cover - trivial availability branch
 except ImportError:  # pragma: no cover
     HAS_NUMPY = False
 
-__version__ = "4.86.0"
+__version__ = "4.87.0"
 
 __all__ = [
     "OptionType",
@@ -2356,6 +2357,9 @@ __all__ = [
     "bessel_jn",
     "bessel_y0",
     "bessel_y1",
+    "fresnel_c",
+    "fresnel_s",
+    "dawson",
     "Dual",
     "dual_derivative",
     "dual_gradient",